Fixes #1 isOnline issue, login field
Reverse logic for isOnline, not sure how this ended up this way, but the
logic was backwards.  It may be that previous versions of Android did
not trigger this correctly.

Force login field to not include spaces.  This has been annoying me for
a while since typing my url hits autocomplete and puts spaces between
words or after periods.